requirements.txt
absl-py==0.9.0
anndata==0.7.4
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
biopython==1.77
bleach==3.1.5
bokeh==2.1.1
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorcet==2.0.2
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
dm-sonnet==1.23
dm-tree==0.1.5
docopt==0.6.2
entrypoints==0.3
forceatlas2-python==1.0
future==0.18.2
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gprofiler-official==1.0.0
grpcio==1.30.0
h5py==2.10.0
harmony-pytorch==0.1.3
holoviews==1.13.3
hvplot==0.6.0
idna==2.9
importlib-metadata==1.7.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.1
Jinja2==2.11.2
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.5
jupyterlab-server==1.1.5
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lightgbm==2.2.1
llvmlite==0.33.0
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
modlamp==4.2.2
mysql-connector-python==8.0.17
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
nose==1.3.7
notebook==6.0.3
numba==0.50.1
numexpr==2.7.1
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
panel==0.9.7
param==1.9.3
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.8.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pyct==0.4.6
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-igraph==0.8.2
pytz==2020.1
pyviz-comms==0.7.6
PyYAML==5.3.1
pyzmq==19.0.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-learn==0.23.1
scikit-misc==0.1.3
scipy==1.4.1
scplot==0.0.16
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
statsmodels==0.11.1
tables==3.6.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
tfp-nightly==0.11.0.dev20200625
threadpoolctl==2.1.0
torch==1.5.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
typing-extensions==3.7.4.2
umap-learn==0.4.6
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.9
zipp==3.1.0
